depth/requirements.txt
chalice==1.17.0
boto3==1.14.43
python-slugify==4.0.1